sdks/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dcp-ai"
version = "1.0.0"
description = "Python SDK for the Digital Citizenship Protocol for AI Agents"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "DCP-AI Contributors" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Security",
]
dependencies = [
    "pynacl>=1.5.0",
    "jsonschema>=4.20.0",
    "pydantic>=2.0",
    "typer>=0.9.0",
    "pqcrypto>=0.4.0",
]

[project.optional-dependencies]
fastapi = ["fastapi>=0.100.0", "uvicorn>=0.24.0"]
langchain = ["langchain>=0.1.0", "langchain-core>=0.1.0"]
openai = ["openai>=1.0.0"]
crewai = ["crewai>=0.1.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21.0"]

[project.scripts]
dcp = "dcp_ai.cli:app"

[project.urls]
Homepage = "https://dcp-ai.dev"
Repository = "https://github.com/dcp-ai-protocol/dcp-ai"
Issues = "https://github.com/dcp-ai-protocol/dcp-ai/issues"

[tool.hatch.build.targets.wheel]
packages = ["dcp_ai"]
